Move pub mod asym_sign out of crypto/mod.rs
into secret/secret_callback.rs.
