+++ spdmlib/src/requester/psk_finish_req.rs
@@ -196,7 +196,7 @@ mod tests_requester {
 
         let pcidoe_transport_encap = &mut PciDoeTransportEncap {};
 
-        crypto::asym_sign::register(ASYM_SIGN_IMPL.clone());
+        crate::secret::asym_sign::register(ASYM_SIGN_IMPL.clone());
         crypto::hmac::register(HMAC_TEST.clone());
 
         let mut responder = responder::ResponderContext::new(
