+++ spdmlib/src/responder/psk_finish_rsp.rs
@@ -173,7 +173,7 @@ mod tests_responder {
         let shared_buffer = SharedBuffer::new();
         let mut socket_io_transport = FakeSpdmDeviceIoReceve::new(&shared_buffer);
 
-        crypto::asym_sign::register(ASYM_SIGN_IMPL.clone());
+        crate::secret::asym_sign::register(ASYM_SIGN_IMPL.clone());
         crypto::hmac::register(HMAC_TEST.clone());
 
         let mut context = responder::ResponderContext::new(
