+++ spdmlib/tests/common/crypto_callbacks.rs
@@ -4,7 +4,7 @@
 
 #![allow(unused)]
 
-use spdmlib::crypto::SpdmAsymSign;
+use spdmlib::secret::SpdmAsymSign;
 
 use spdmlib::protocol::{
     SpdmBaseAsymAlgo, SpdmBaseHashAlgo, SpdmSignatureStruct, RSAPSS_2048_KEY_SIZE,
